# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: RLEGS AM Monitoring & Bot Telegram API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /telegram/send:
    post:
      operationId: sendTelegram
      tags: [telegram]
      summary: Send Telegram message to AMs
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendTelegramBody"
      responses:
        "200":
          description: Send result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramSendResult"
...
  /telegram/register-code:
    post:
      operationId: generateTelegramCode
      tags: [telegram]
      summary: Generate Telegram registration code for an AM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateTelegramCodeBody"
      responses:
        "200":
          description: Code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TelegramCodeResponse"
...
components:
  schemas:
...
    SendTelegramBody:
      type: object
      properties:
        targetNiks:
          type: ["array", "null"]
          items:
            type: string
        period:
          type: string
        includePerformance:
          type: boolean
        includeFunnel:
          type: boolean
        includeActivity:
          type: boolean
        customMessage:
          type: ["string", "null"]
      required: [period, includePerformance, includeFunnel, includeActivity]
    TelegramSendResult:
      type: object
      properties:
        sent:
          type: integer
        failed:
          type: integer
        skipped:
          type: integer
        details:
          type: array
          items:
            $ref: "#/components/schemas/TelegramSendDetail"
      required: [sent, failed, skipped, details]
    TelegramSendDetail:
      type: object
      properties:
        nik:
          type: string
        namaAm:
          type: string
        status:
          type: string
        error:
          type: ["string", "null"]
      required: [nik, namaAm, status]
...
    GenerateTelegramCodeBody:
      type: object
      properties:
        amId:
          type: integer
      required: [amId]
    TelegramCodeResponse:
      type: object
      properties:
        code:
          type: string
        expiresAt:
          type: string
      required: [code, expiresAt]
...